Move verify_checksum() from general_js to general
This utility function will also be useful for Pip and potentially other
package managers.
