+++ tests/test_workers/test_pkg_managers/test_general.py
@@ -9,6 +9,8 @@ from cachito.workers.pkg_managers.general import (
     update_request_with_config_files,
     update_request_with_deps,
     update_request_with_package,
+    verify_checksum,
+    ChecksumInfo,
 )
 
 
@@ -119,3 +121,37 @@ def test_update_request_with_config_files_failed(mock_requests):
     expected = "Adding configuration files on request 1 failed"
     with pytest.raises(CachitoError, match=expected):
         update_request_with_config_files(1, [])
+
+
+def test_verify_checksum(tmpdir):
+    file = tmpdir.join("spells.txt")
+    file.write("Beetlejuice! Beetlejuice! Beetlejuice!")
+
+    expected = {
+        "sha512": (
+            "da518fe8b800b3325fe35ca680085fe37626414d0916937a01a25ef8f5d7aa769b7233073235fce85ee"
+            "c717e02bb9d72062656cf2d79223792a784910c267b54"
+        ),
+        "sha256": "ed1f8cd69bfacf0528744b6a7084f36e8841b6128de0217503e215612a0ee835",
+        "md5": "308764bc995153f7d853827a675e6731",
+    }
+    for algorithm, checksum in expected.items():
+        verify_checksum(str(file), ChecksumInfo(algorithm, checksum))
+
+
+def test_verify_checksum_invalid_hexdigest(tmpdir):
+    file = tmpdir.join("spells.txt")
+    file.write("Beetlejuice! Beetlejuice! Beetlejuice!")
+
+    expected_error = "The file spells.txt has an unexpected checksum value"
+    with pytest.raises(CachitoError, match=expected_error):
+        verify_checksum(str(file), ChecksumInfo("sha512", "spam"))
+
+
+def test_verify_checksum_unsupported_algorithm(tmpdir):
+    file = tmpdir.join("spells.txt")
+    file.write("Beetlejuice! Beetlejuice! Beetlejuice!")
+
+    expected_error = "Cannot perform checksum on the file spells.txt,.*bacon.*"
+    with pytest.raises(CachitoError, match=expected_error):
+        verify_checksum(str(file), ChecksumInfo("bacon", "spam"))
